refactoring: Introduce separate PCS module
