+++ benches/pcs.rs
@@ -1,4 +1,4 @@
-use arecibo::provider::{
+use arecibo::provider::pcs::{
   hyperkzg::EvaluationEngine as MLEvaluationEngine,
   ipa_pc::EvaluationEngine as IPAEvaluationEngine, non_hiding_zeromorph::ZMPCS,
 };
