+++ examples/hashchain.rs
@@ -27,8 +27,8 @@ use std::time::Instant;
 
 type E1 = Bn256EngineKZG;
 type E2 = GrumpkinEngine;
-type EE1 = arecibo::provider::hyperkzg::EvaluationEngine<Bn256, E1>;
-type EE2 = arecibo::provider::ipa_pc::EvaluationEngine<E2>;
+type EE1 = arecibo::provider::pcs::hyperkzg::EvaluationEngine<Bn256, E1>;
+type EE2 = arecibo::provider::pcs::ipa_pc::EvaluationEngine<E2>;
 type S1 = arecibo::spartan::snark::RelaxedR1CSSNARK<E1, EE1>; // non-preprocessing SNARK
 type S2 = arecibo::spartan::snark::RelaxedR1CSSNARK<E2, EE2>; // non-preprocessing SNARK
