+++ src/lib.rs
@@ -224,7 +224,7 @@ where
   ///
   /// ```rust
   /// # use arecibo::spartan::ppsnark::RelaxedR1CSSNARK;
-  /// # use arecibo::provider::ipa_pc::EvaluationEngine;
+  /// # use arecibo::provider::pcs::ipa_pc::EvaluationEngine;
   /// # use arecibo::provider::{PallasEngine, VestaEngine};
   /// # use arecibo::traits::{circuit::TrivialCircuit, Engine, snark::RelaxedR1CSSNARKTrait};
   /// use arecibo::PublicParams;
@@ -1022,8 +1022,8 @@ mod tests {
   use super::*;
   use crate::{
     provider::{
-      non_hiding_zeromorph::ZMPCS, Bn256EngineIPA, Bn256EngineKZG, Bn256EngineZM, PallasEngine,
-      Secp256k1Engine,
+      pcs, pcs::non_hiding_zeromorph::ZMPCS, Bn256EngineIPA, Bn256EngineKZG, Bn256EngineZM,
+      PallasEngine, Secp256k1Engine,
     },
     traits::{evaluation::EvaluationEngineTrait, snark::default_ck_hint},
   };
@@ -1034,7 +1034,7 @@ mod tests {
   use halo2curves::bn256::Bn256;
   use traits::circuit::TrivialCircuit;
 
-  type EE<E> = provider::ipa_pc::EvaluationEngine<E>;
+  type EE<E> = pcs::ipa_pc::EvaluationEngine<E>;
   type S<E, EE> = spartan::snark::RelaxedR1CSSNARK<E, EE>;
   type SPrime<E, EE> = spartan::ppsnark::RelaxedR1CSSNARK<E, EE>;
 
@@ -1375,7 +1375,7 @@ mod tests {
     test_ivc_nontrivial_with_compression_with::<Bn256EngineZM, ZMPCS<Bn256, _>, EE<_>>();
     test_ivc_nontrivial_with_compression_with::<
       Bn256EngineKZG,
-      provider::hyperkzg::EvaluationEngine<Bn256, _>,
+      pcs::hyperkzg::EvaluationEngine<Bn256, _>,
       EE<_>,
     >();
   }
@@ -1400,7 +1400,7 @@ mod tests {
     test_ivc_nontrivial_with_spark_compression_with::<Bn256EngineZM, ZMPCS<Bn256, _>, EE<_>>();
     test_ivc_nontrivial_with_spark_compression_with::<
       Bn256EngineKZG,
-      provider::hyperkzg::EvaluationEngine<Bn256, _>,
+      pcs::hyperkzg::EvaluationEngine<Bn256, _>,
       EE<_>,
     >();
   }
@@ -1429,7 +1429,7 @@ mod tests {
     test_ivc_nontrivial_with_batched_compression_with::<Bn256EngineZM, ZMPCS<Bn256, _>, EE<_>>();
     test_ivc_nontrivial_with_batched_compression_with::<
       Bn256EngineKZG,
-      provider::hyperkzg::EvaluationEngine<Bn256, _>,
+      pcs::hyperkzg::EvaluationEngine<Bn256, _>,
       EE<_>,
     >();
   }
@@ -1457,7 +1457,7 @@ mod tests {
     );
     test_ivc_nontrivial_with_batched_spark_compression_with::<
       Bn256EngineKZG,
-      provider::hyperkzg::EvaluationEngine<Bn256, _>,
+      pcs::hyperkzg::EvaluationEngine<Bn256, _>,
       EE<_>,
     >();
   }
