+++ src/provider/pcs/mod.rs
@@ -1,0 +1,9 @@
+/// TODO write docs for PCS module
+///
+// public modules to be used as an evaluation engine with Spartan
+pub mod hyperkzg;
+pub mod ipa_pc;
+// a non-hiding variant of {kzg, zeromorph}
+pub mod kzg_commitment;
+pub mod non_hiding_zeromorph;
+pub mod pedersen;
